feat!: Add configureLogger() to /ex module
