+++ .changeset/fuzzy-words-scream.md
@@ -1,0 +1,5 @@
+---
+'@collagejs/vite-css': minor
+---
+
+feat!(css): Add configureLogger() to /ex module
