+++ packages/css/src/ex-defs.ts
@@ -10,3 +10,11 @@ export const cssHelpersModuleName = './css-helpers.js';
  * Module name of the type definitions module.
  */
 export const typesModuleName = './ex-types.js';
+/**
+ * Module name of the logger module.
+ */
+export const cssLoggerModuleName = './logger.js';
+/**
+ * Array of all module names that are dynamically inserted by this plug-in for Vite projects to provide CSS injection.
+ */
+export const allModuleNames = [extensionModuleName, cssHelpersModuleName, cssLoggerModuleName];
