+++ packages/css/src/index-ex.ts
@@ -1,5 +1,6 @@
 import type { ViteEnv } from "./ex-types.js";
 export { cssMountFactory } from "./css.js";
 export type * from "./ex-types.js";
+export { configureLogger } from "./logger.js";
 
 export const viteEnv = {} as ViteEnv;
